pyproject.toml
[project]
name = "threatcorrelator"
version = "0.1.0"
description = "A modular threat intelligence tool for fetching and correlating IOCs"
authors = [
    { name = "Alexander Zimpher", email = "alexanderzimpher@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "click>=8.2.1,<9.0.0",
    "requests>=2.32.3,<3.0.0",
    "pyyaml>=6.0.2,<7.0.0",
    "sqlalchemy>=2.0.0,<3.0.0",
    "otxv2 (>=1.5.12,<2.0.0)",
    "pdfkit>=1.0.0,<2.0.0",
    "reportlab>=4.0.0,<5.0.0",
    "geoip2>=4.8.0,<5.0.0",
    "ipwhois>=1.2.0,<2.0.0"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{ include = "threatcorrelator", from = "src" }]

[tool.poetry.scripts]
threatcorrelator = "threatcorrelator.cli:main"

[tool.poetry.group.dev.dependencies]
streamlit = "^1.45.1"
pandas = "^2.2.3"
plotly = "^6.1.2"
pytest = "^8.3.5"
pytest-mock = "^3.14.1"
flake8 = "^7.2.0"
pytest-cov = "^5.0.0"
bandit = "^1.8.5"
